Refactor retal management
